Converted InitCommPixel to a uint UIntPixel.  mapcode is now consistently uint.
git-svn-id: http://Marc-PC/svn/Full@258 85f15e0e-5137-b144-a863-98bd542365a3

+++ trunk/succession-library/trunk/src/initial-communities/ICommunity.cs
@@ -11,7 +11,7 @@ namespace Landis.Library.InitialCommunities
         /// <summary>
         /// The code that represents the community on maps.
         /// </summary>
-        ushort MapCode
+        uint MapCode
         {
             get;
         }
